processor/requirements.txt
numpy==1.19.5
psycopg2-binary==2.8.6
requests==2.25.1
spacy==2.3.5
wmd==1.3.2
pytest==6.2.1
pytest-mock==3.5.1
